EKYC: refactored for merging
