+++ lib/core/navigation/models/ekyc_event_data.dart
