+++ lib/core/network/managers/signalr_connection_manager.dart
@@ -11,6 +11,7 @@
  */
 
 import 'dart:convert';
+import 'dart:developer';
 
 import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart' show kDebugMode;
@@ -74,7 +75,7 @@ class SignalrConnectionManager {
   void listenForTransitionEvents({required Function(NeoSignalRTransition transition) onTransition}) {
     _hubConnection?.on(methodName, (List<Object?>? transitions) {
       if (kDebugMode) {
-        debugPrint('\n[SignalrConnectionManager] Transition: $transitions');
+        log('\n[SignalrConnectionManager] Transition: $transitions');
       }
       if (transitions == null) {
         return;
